tested wildcard_handle with many inputs. splitted across many files to fit in the 5 fucntions per file rule. test_wildcard.sh isnt needed anymore it will be tested by system test
