+++ wildcard_handle/wildcard_handle.c
@@ -10,25 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../main/minishell.h"
-#include <dirent.h>
+#include "wildcard_handler.h"
 
-void print_result(t_split *test)
-{
-    int i;
-    printf("\n\tsize: %d", test->size);
-    printf("\n\tstart:%3p", test->start);
-    i = 0;
-    while(i <= test->size)
-    {
-        printf("\n\t\t%-2d:%s", i, test->start[i]);
-        i++;
-    }
-    printf(" -> after last value\n");
-}
-
-
-int	count_dir_entries(const char *directory)
+static int	count_dir_entries(const char *directory)
 {
 	DIR				*dir;
 	struct dirent	*entry;
@@ -46,7 +30,7 @@ int	count_dir_entries(const char *directory)
 	return (count);
 }
 
-t_split	get_cur_dir_entries(const char *direc)
+static t_split	get_cur_dir_entries(const char *direc)
 {
 	t_split			cur_dir_ent;
 	DIR				*dir;
@@ -70,193 +54,7 @@ t_split	get_cur_dir_entries(const char *direc)
 	return (cur_dir_ent);
 }
 
-t_split	create_filter(const char *wildcard)
-{
-	t_split		filter;
-	int			indx[3];
-	char 		quote;
-
-	quote  =  0;
-	ft_memset(indx, 0, sizeof(int) * 3);
-	filter.size = countchr_not_quote((char *)wildcard, '*') + 1;
-	filter.start = (char **)ft_calloc(filter.size + 1, sizeof(char *));
-	if (filter.start == NULL)
-		return (create_split(NULL, 0));
-	while (indx[0] <= ft_strlen(wildcard))
-	{
-		while (wildcard[indx[1]] != '\0' && wildcard[indx[1]] != '*')
-			indx[1]++;
-		if (indx[0] != indx[1])
-			filter.start[indx[2]++] = ft_substr(wildcard, indx[0], indx[1] - indx[0]);
-		else
-			filter.start[indx[2]++] = ft_strdup("");
-		indx[0] = indx[1] + 1;
-		indx[1] = indx[0];
-	}
-	return (filter);
-}
-
-void apply_filter_start(t_split filter, t_split cur_dir, char *check_list)
-{
-	int i;
-
-	i = 0;
-	if (filter.start[0][0] == '\0')
-	{
-		while (i < cur_dir.size)
-		{
-			if (cur_dir.start[i][0] == '.')
-				check_list[i] = '0';
-			i++;
-		}
-	}
-	else
-	{
-		while (i < cur_dir.size)
-		{
-			if (ft_strncmp(filter.start[0], cur_dir.start[i], ft_strlen(filter.start[0])))
-				check_list[i] = '0';
-			i++;
-		}
-	}
-}
-
-int ft_strrcmp(const char *s1, const char *s2)
-{
-	int s1i;
-	int s2i;
-
-	s1i = ft_strlen(s1) - 1;
-	s2i = ft_strlen(s2) - 1;
-	if (s2i < s1i)
-		return (1);
-	while (s1i > 0)
-	{
-		if (s1[s1i] != s2[s2i])
-			return (s1[s1i] - s2[s2i]);
-		s2i--;
-		s1i--;
-	}
-	return (s1[0] - s2[s2i]);
-}
-
-void apply_filter_end(t_split filter, t_split cur_dir, char *check_list)
-{
-	int i;
-	char *filter_end;
-
-	filter_end = filter.start[filter.size - 1];
-	i = 0;
-	while (i < cur_dir.size && check_list[i] == '0')
-				i++;
-	if (filter.start[filter.size - 1][0] != '\0')
-		while (i < cur_dir.size)
-		{
-			if (ft_strrcmp(filter_end, cur_dir.start[i]))
-				check_list[i] = '0';
-			i++;
-			while (check_list[i] == '0')
-				i++;
-		}
-}
-
-void apply_filter_minlen(t_split filter, t_split cur_dir, char *check_list, char *wildcard)
-{
-	int i;
-	char *filter_end;
-	int minlen;
-
-	i = 0;
-	minlen = ft_strlen(wildcard) - filter.size + 1;
-	while (i < cur_dir.size)
-	{
-		if (ft_strlen(cur_dir.start[i]) < minlen)
-			check_list[i] = '0';
-		i++;
-	}
-}
-
-void apply_filter_midsin(t_split filter, char *cur_dir_i, char *check_list_i)
-{
-	int filter_i;
-	char *index;
-
-	filter_i = 1;
-	index = &(cur_dir_i[ft_strlen(filter.start[0])]);
-	while (filter_i < filter.size - 1)
-	{
-		index = (ft_strnstr(index, filter.start[filter_i], ft_strlen(index) \
-		- ft_strlen(filter.start[filter.size - 1])));
-		if (index == NULL)
-		{
-			*check_list_i = '0';
-			return ;
-		}
-		if (ft_strlen(index) > ft_strlen(filter.start[filter_i]))
-			index = &(index[ft_strlen(filter.start[filter_i])]);
-		else
-			index = &(index[ft_strlen(index)]);
-		filter_i++;
-	}
-}
-
-void apply_filter_middle(t_split filter, t_split cur_dir, char *check_list)
-{
-	int i;
-	char *filter_middle;
-	
-	i = 0;
-	if (filter.size < 3)
-		return ;
-	while (i < cur_dir.size)
-	{
-		if (check_list[i] != '0')
-			apply_filter_midsin(filter, cur_dir.start[i], &(check_list[i]));
-		i++;
-	}
-}
-void apply_filter(t_split cur_dir, char *check_list, char *wildcard)
-{
-	t_split filter;
-
-	filter = create_filter(wildcard);
-	apply_filter_minlen(filter, cur_dir, check_list, wildcard);
-	apply_filter_start(filter, cur_dir, check_list);
-	apply_filter_end(filter, cur_dir, check_list);
-	apply_filter_middle(filter, cur_dir, check_list);
-	free_split(&filter);
-}
-
-char *ft_strjoin_sq_f(char **to_free, const char *to_add)
-{
-	char *ret;
-	int ret_size;
-
-	
-	if (*to_free == NULL)
-	{
-		ret_size = ft_strlen(to_add) + 3;
-		ret = (char *)ft_calloc(ret_size, sizeof(char));
-		if (ret == NULL)
-			return (NULL);
-		ft_strlcpy(ret, "'", ret_size);
-	}
-	else
-	{
-		ret_size = ft_strlen(*to_free) + ft_strlen(to_add) + 4;
-		ret = (char *)ft_calloc(ret_size, sizeof(char));
-		if (ret == NULL)
-			return (NULL);
-		ft_strlcpy(ret, *to_free, ret_size);
-		ft_strlcat(ret, " '", ret_size);
-		free(*to_free);
-	}
-	ft_strlcat(ret, to_add, ret_size);
-	ft_strlcat(ret, "'", ret_size);
-	return (ret);
-}
-
-char	*wildcard_handle(char *wildcard)
+static char	*wildcard_handle(char *wildcard)
 {
 	t_split	cur_dir;
 	char *check_list;
@@ -310,7 +108,7 @@ char *wildcard_input_modify(char *current_input)
 }
 
 
-//cc wildcard_handle/wildcard_handle.c e-libft/libft.a and_or_parser/and_or_utils.c  t_split_utils/t_split_utils.c t_split_utils/t_split_utils2.c t_split_utils/ft_split_quotes.c -o test
+//cc wildcard_handle/wildcard_handle.c wildcard_handle/wildcard_filter.c wildcard_handle/wildcard_filter2.c wildcard_handle/wildcard_utils.c  e-libft/libft.a and_or_parser/and_or_utils.c  t_split_utils/t_split_utils.c t_split_utils/t_split_utils2.c t_split_utils/ft_split_quotes.c
 int	main(int ac, char **av)
 {	
 	char *a = wildcard_input_modify(av[1]);
