+++ wildcard_handle/wildcard_handler.h
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   wildcard_handler.h                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kuzyilma <kuzyilma@student.42istanbul.c    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/01 13:53:36 by kuzyilma          #+#    #+#             */
+/*   Updated: 2025/06/01 14:12:06 by kuzyilma         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../main/minishell.h"
+#include <dirent.h>
+
+void apply_filter_minlen(t_split filter, t_split cur_dir, char *check_list, char *wildcard);
+void apply_filter_start(t_split filter, t_split cur_dir, char *check_list);
+void apply_filter_middle(t_split filter, t_split cur_dir, char *check_list);
+void apply_filter_end(t_split filter, t_split cur_dir, char *check_list);
+void apply_filter(t_split cur_dir, char *check_list, char *wildcard);
+char *wildcard_input_modify(char *current_input);
+int ft_strrcmp(const char *s1, const char *s2);
+char *ft_strjoin_sq_f(char **to_free, const char *to_add);
